requirements.txt
# ─── Web Framework ─────────────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.9       # File upload support ke liye ZARURI
aiofiles==23.2.1               # Async file operations

# ─── AI / ML ────────────────────────────────────────────────────────────────
# NOTE: torch CPU-only version Docker image chota rakhne ke liye
# GPU ke liye: torch==2.1.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0
# openai-whisper GitHub se install hoga (Dockerfile mein)

# ─── Database ───────────────────────────────────────────────────────────────
sqlalchemy==2.0.25
# PostgreSQL ke liye uncomment karo:
# psycopg2-binary==2.9.9

# ─── Utilities ──────────────────────────────────────────────────────────────
python-dotenv==1.0.0
